requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.20.0
pyyaml>=6.0
click>=8.0.0

# Data processing
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# AI Integration
openai>=1.0.0
azure-identity>=1.15.0

# Schema parsing and database utilities
sqlparse>=0.4.0
lxml>=4.9.0

# Logging and configuration
python-dotenv>=0.19.0

# Text processing and NLP
nltk>=3.8.0
textblob>=0.17.0

# Visualization (optional for advanced reporting)
matplotlib>=3.5.0
seaborn>=0.11.0
